main delegates to --existing or --explore
