+++ experiments/lmdb-oxen/src/existing/migrate.rs
@@ -1,8 +1,8 @@
 use std::collections::HashSet;
 use std::path::PathBuf;
 
-use crate::framework::FrameworkResult;
-use crate::lmdb;
+use super::framework::FrameworkResult;
+use super::lmdb;
 
 const LMDB_PATH: &str = "./.oxen/tree/lmdb";
 use liboxen::core::v_latest::index::CommitMerkleTree;
