+++ experiments/lmdb-oxen/src/explore.rs
@@ -1,0 +1,3 @@
+pub mod new_main;
+// scratch.rs contains design notes / pseudo-code, not compilable Rust
+// pub mod scratch;
